hokimiyat-master/requirements.txt
telethon==1.42.0
openai==2.12.0
python-dotenv==1.2.1
asyncio==4.0.0